refactor: Dto 필드 형식 변경 (#283)

+++ src/main/java/org/sopt/app/application/auth/PlaygroundAuthInfo.java
@@ -184,6 +184,6 @@ public class PlaygroundAuthInfo {
     @Getter
     public static class PlaygroundActivity {
         private String part;
-        private Long generation;
+        private Integer generation;
     }
 }
